+++ _bangle-website/docs/examples/images.example1.jsx
@@ -1,11 +1,11 @@
 import '@bangle.dev/core/style.css';
 
-import { BangleEditor, BangleEditorState, image } from '@bangle.dev/core';
+import { BangleEditor, BangleEditorState, components } from '@bangle.dev/core';
 
 export default function Editor(domNode) {
   const state = new BangleEditorState({
-    specs: [image.spec()],
-    plugins: () => [image.plugins()],
+    specs: [components.image.spec()],
+    plugins: () => [components.image.plugins()],
     initialValue: `<div><p>Hey there!</p>
         <p>Let us see an image of cat</p>
         <img src="https://media.giphy.com/media/7xkxbhryQO7hm/giphy.gif" />