+++ core/index.js
@@ -1,12 +1,15 @@
 import * as logging from './utils/logging';
-import browser, { isChromeWithSelectionBug } from './utils/browser';
+import browser from './utils/browser';
 import * as utils from './utils/utils';
+import * as components from './components/components';
 
-export * from './components/components';
 export * from './bangle-editor';
 export * from './bangle-editor-state';
 export * from './node-view';
 export * from './spec-registry';
 export * from './plugin';
 
-export { utils, logging, browser, isChromeWithSelectionBug };
+const isChromeWithSelectionBug =
+  browser.chrome && !browser.android && browser.chrome_version >= 58;
+
+export { components, utils, logging, browser, isChromeWithSelectionBug };